IPL Explorer/backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.135.2
uvicorn[standard]==0.34.0
gunicorn==23.0.0

# ── Data ──────────────────────────────────────────────────────────────────────
pandas==2.2.3
pyyaml==6.0.2

# ── AI ────────────────────────────────────────────────────────────────────────
anthropic==0.87.0

# ── Validation & config ───────────────────────────────────────────────────────
pydantic==2.11.1
python-dotenv==1.1.0

# ── HTTP client ───────────────────────────────────────────────────────────────
httpx==0.28.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.5
pytest-asyncio==0.25.3

# ── Machine Learning ─────────────────────────────────────────────────────────
scikit-learn==1.6.1       # used only in train_win_model.py (pipeline), not at runtime
# Note: the inference service (ml_service.py) uses pure numpy — sklearn not needed
# for the running backend, only for re-training the model
